requirements.txt
matplotlib==3.7.2
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.1
streamlit==1.25.0
